Update security.html
Added a link to https://cryptosec.info/exit-scams/

It's a decently long list of cryptocurrency exit scams in general. While it's for cryptocurrencies in general and not solely concerning bitcoin, the list also contains bitcoin/crypto exchanges so It'll probably be a decent warning for people to stop leaving funds on exchanges.

Hope you like it.